Improve readability [1]
